listDelete _settings refactoring

+++ listDelete.py
@@ -1,26 +1,31 @@
 import json
 import os
-from _settings import listsFile, fileDir
+from _settings import listsFile, fileDir, commonResult
 
 
 def listDelete(listId = -1):
     """Удаление списка"""
 
     fileName = fileDir + str(listId)+'.json'
-    if os.path.exists(fileName):
-        print(fileName)
-        os.remove(fileName)
+    result = commonResult
 
     with open(listsFile, 'r', encoding='utf-8') as f:
         listData = json.load(f)
 
-    listData = list(filter(lambda p: p['id'] != listId, listData))
+    isListExists = len(list(filter(lambda p: p['id'] == listId, listData))) > 0
+    if isListExists:
+        listData = list(filter(lambda p: p['id'] != listId, listData))
+        resultJson = json.dumps(listData, ensure_ascii=False, indent=2)
 
-    resultJson = json.dumps(listData, ensure_ascii=False, indent=2)
+        with open(listsFile, 'w', encoding='utf-8') as f:
+            f.write(resultJson)
 
-    with open(listsFile, 'w', encoding='utf-8') as f:
-        f.write(resultJson)
+        """Удаление файла с содержимым удаляемого списка"""
+        if os.path.exists(fileName):
+            os.remove(fileName)
+    else:
+        result = dict(isError=True, errorText=f'list id {listId} is not exists')
 
-    return
+    return result
 
-# listDelete(99)
+# print(listDelete(4))
